:ambulance:: Fixed docker errors Renamed `server` to `bg_api`.
