+++ nginx/default.conf
@@ -1,23 +1,23 @@
-upstream client {
-    server client:5000;
-}
+# upstream client {
+#     bg_api client:5000;
+# }
 
 upstream api {
     server api:5000;
 }
 
-upstream godmod {
+upstream godmode {
     server adminer:8080;
 }
 
 server {
-    listen 80 http2;
+    listen 80;
 
     #server_name ctf-444.ru;
     keepalive_timeout 70;
 
     location / {
-        proxy_pass http://client;
+        proxy_pass http://api;
     }
 
     location /api {
