--- requirements.txt
@@ -1,6 +1,0 @@
-Flask==2.2.3
-bcrypt==4.0.1
-SQLAlchemy==2.0.9
-python-dotenv==1.0.0
-PyJWT==2.6.0
-Werkzeug==2.2.3
